+++ electron-builder.js
@@ -1,5 +1,4 @@
 const {execSync} = require('child_process')
-const path = require('path')
 const pkg = require('./package.json')
 
 const sh = (cmd) => {
@@ -37,5 +36,21 @@ module.exports = {
   mac: {
     category: 'public.app-category.developer-tools',
     target: ['dmg'],
+  },
+
+  dmg: {
+    iconSize: 160,
+    contents: [
+      {
+        x: 180,
+        y: 170,
+      },
+      {
+        x: 480,
+        y: 170,
+        type: 'link',
+        path: '/Applications',
+      },
+    ],
   },
 }