+++ src/page/library-rule-list/ConfigEditor.js
@@ -7,14 +7,13 @@ import React, {
   useImperativeHandle,
   forwardRef,
 } from 'react'
-import {usePersistFn, useWhyDidYouUpdate, useUpdateEffect} from 'ahooks'
-
+import {usePersistFn, useUpdateEffect} from 'ahooks'
+import {Spin} from 'antd'
 import style from './ConfigEditor.module.less'
 
 export default forwardRef(ConfigEditor)
-
 function ConfigEditor(props, ref) {
-  const {value, onChange, readonly} = props
+  const {value, onChange, readonly, spinProps} = props
 
   const editorWillMount = usePersistFn((editor, monaco) => {})
 
@@ -36,7 +35,7 @@ function ConfigEditor(props, ref) {
   })
 
   const editorOnChange = usePersistFn((newValue, e) => {
-    props.onChange(newValue)
+    onChange(newValue)
   })
 
   const options = {
@@ -59,16 +58,18 @@ function ConfigEditor(props, ref) {
   }, [readonly, ref])
 
   return (
-    <div style={{width: '100%', height: '50vh'}} className={style.editor}>
-      <MonacoEditor
-        language='yaml'
-        theme='quite-light'
-        value={value}
-        options={options}
-        onChange={editorOnChange}
-        editorWillMount={editorWillMount}
-        editorDidMount={editorDidMount}
-      />
-    </div>
+    <Spin {...spinProps}>
+      <div style={{width: '100%', height: '50vh'}} className={style.editor}>
+        <MonacoEditor
+          language='yaml'
+          theme='quite-light'
+          value={value}
+          options={options}
+          onChange={editorOnChange}
+          editorWillMount={editorWillMount}
+          editorDidMount={editorDidMount}
+        />
+      </div>
+    </Spin>
   )
 }