+++ src/page/library-rule-list/model/index.js
@@ -14,6 +14,12 @@ export default {
     list: [],
   },
 
+  listen: {
+    'global/reload'() {
+      this.load()
+    },
+  },
+
   effects: (dispatch) => {
     return {
       load() {