+++ src/store.js
@@ -1,6 +1,7 @@
 import {init} from '@rematch/core'
 import immerPlugin from '@rematch/immer'
 import statePlugin from '@magicdawn/x/rematch/state-plugin'
+import listenPlugin from './util/rematch-listen-plugin'
 
 import * as models from './models'
 import librarySubscribe from './page/library-subscribe/model/index'
@@ -10,7 +11,7 @@ import preference from './page/preference/model/index'
 
 const store = init({
   models: {...models, librarySubscribe, libraryRuleList, currentConfig, preference},
-  plugins: [immerPlugin(), statePlugin()],
+  plugins: [immerPlugin(), statePlugin(), listenPlugin()],
 })
 
 export default store