+++ src/util/fn/gen.js
@@ -1,9 +1,9 @@
 const {homedir} = require('os')
 const {join: pathjoin} = require('path')
 const fse = require('fs-extra')
-const _ = require('lodash')
 const Yaml = require('js-yaml')
 import store from '../../store'
+import request from 'umi-request'
 
 export default async function genConfig() {
   const rootState = store.getState()
@@ -49,31 +49,35 @@ export default async function genConfig() {
   for (let r of ruleArr) {
     const {item} = r
     const {type, url, content} = item
+    let usingContent = content
 
     // remote: url -> content
     if (type === 'remote') {
-      // TODO: 支持 remote
+      usingContent = await request.get(url, {
+        responseType: 'text',
+        headers: {'x-extra-headers': JSON.stringify({'user-agent': 'clash'})},
+      })
     }
 
-    // local: use content
-    if (type === 'local') {
-      const cur = Yaml.load(content)
-      const {Rule, ...otherConfig} = cur
-      config = {...otherConfig, ...config, Rule: [...(config.Rule || []), ...(Rule || [])]}
-    }
+    const cur = Yaml.load(usingContent)
+    const {rules, ...otherConfig} = cur
+    config = {...otherConfig, ...config, rules: [...(config.rules || []), ...(rules || [])]}
   }
 
   /**
    * subscribe
    */
 
-  if (!config['Proxy'] || !Array.isArray(config['Proxy'])) {
-    config['Proxy'] = []
+  if (!config.proxies || !Array.isArray(config.proxies)) {
+    config.proxies = []
+  }
+  if (!config['proxy-groups'] || !Array.isArray(config['proxy-groups'])) {
+    config['proxy-groups'] = []
   }
 
   for (let s of subscribeArr) {
     const {item} = s
-    const {name, url} = item
+    const {url} = item
     let servers
 
     // update subscribe
@@ -87,11 +91,11 @@ export default async function genConfig() {
       servers = subscribeDetail[url]
     }
 
-    config['Proxy'] = config['Proxy'].concat(servers)
+    config.proxies = config.proxies.concat(servers)
   }
 
-  const proxyGroups = config['Proxy Group']
-  const allProxies = config['Proxy'].map((row) => row.name)
+  const proxyGroups = config['proxy-groups']
+  const allProxies = config.proxies.map((row) => row.name)
   proxyGroups
     .filter((item) => !item.proxies || !Array.isArray(item.proxies) || !item.proxies.length)
     .forEach((item) => {
@@ -99,10 +103,10 @@ export default async function genConfig() {
     })
 
   const existNames = proxyGroups.map((i) => i.name)
-  for (let line of config.Rule) {
+  for (let line of config.rules) {
     const use = line.split(',').slice(-1)[0]
     if (!use) continue
-    if (['DIRECT', 'REJECT'].includes(use)) continue
+    if (['DIRECT', 'REJECT', 'no-resolve'].includes(use)) continue
     if (existNames.includes(use)) continue
 
     // - {name: Others, type: select, proxies: [DIRECT, Proxy]}
@@ -111,7 +115,7 @@ export default async function genConfig() {
       type: 'select',
       proxies: ['DIRECT', 'Proxy', 'REJECT'],
     }
-    config['Proxy Group'].push(newgroup)
+    config['proxy-groups'].push(newgroup)
     existNames.push(use)
   }
 
@@ -124,19 +128,6 @@ export default async function genConfig() {
     success: true,
     msg: `${file} writed`,
   }
-
-  // const ruleByGithubContent = await readUrl({
-  //   // url: 'https://raw.githubusercontent.com/lhie1/Rules/master/Clash/Rule.yml',
-  //   https://raw.githubusercontent.com/lhie1/Rules/master/Clash/Rule.yaml
-  //   // url: 'https://cdn.jsdelivr.net/gh/lhie1/Rules@master/Clash/Rule.yaml',
-  //   // url: 'https://raw.githubusercontent.com/lhie1/Rules/master/Clash/Rule.yaml',
-  //   url: 'https://cdn.jsdelivr.net/gh/lhie1/Rules@master/Clash/Rule.yaml',
-  //   encoding: 'utf8',
-  // })
-  // const ruleByGithubContent = fse.readFileSync(__dirname + '/../remote/Rule.yml', 'utf8')
-  // const ruleByGithub = {
-  //   Rule: Yaml.load(ruleByGithubContent),
-  // }
 }
 
 export const DEFAULT_NAME = 'clash-config-manager'