+++ CHANGELOG.md
@@ -1,5 +1,31 @@
 # Changelog
 
+## v0.0.7 2020-09-19
+
+- support remote config file
+
+## v0.0.6 2020-09-19
+
+- 适配 clash core 1.0, see https://github.com/Dreamacro/clash/wiki/breaking-changes-in-1.0.0
+
+## v0.0.5 2020-09-19
+
+- fix error can not find command `atom` / `code`
+
+## v0.0.4 2020-09-19
+
+- [x] 快速添加规则, mc clash add-rule GUI version
+- [x] 记住窗口位置
+- [x] 导入导出(store 加密有必要, 防止扫描)
+- [ ] 备份不处理详情. (no need)
+- [x] 在 vscode/Atom 中编辑规则
+
+## v0.0.3 2020-09-19
+
+- fix dmg icon
+- feat add rule
+- ect...
+
 ## v0.0.2 2020-09-18
 
 - add icons & make modals centered, etc UI modifications.